requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
sqlmodel==0.0.16

# Auth & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.122.0

# AI
openai==1.14.3

# Notifications
firebase-admin==6.5.0

# AWS
boto3==1.34.80

# Other
httpx==0.27.0

# Dev & Linting
pytest==8.1.1
pytest-cov==4.1.0
black==24.4.0
flake8==7.0.0
pre-commit==3.7.0
requests==2.31.0
Faker==25.2.0
fastapi-mail==1.4.1
